Rework API to use AST transforms on mod classes
This also ensures that static compilation works properly, instead of
having to rely on a mix of dynamic and static compilation. On the other
side, this also causes a small binary break due to `EventBus` being
moved to its own class, which should not really cause issues at this
stage of development, mostly because 1.0.0 had additional issues that
made using the annotation problematic in certain contexts.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>